Main/requirements.txt
msal
requests
python-dotenv
anthropic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client


